Remove `weak_object_base.hpp`, forward decl is enough to break cycle
