+++ sycl/include/sycl/accessor.hpp
@@ -24,7 +24,6 @@
 #include <sycl/detail/type_traits.hpp>                // for const_if_const_AS
 #include <sycl/exception.hpp>                         // for make_error_code
 #include <sycl/ext/oneapi/accessor_property_list.hpp> // for accessor_prope...
-#include <sycl/ext/oneapi/weak_object_base.hpp>       // for getSyclWeakObj...
 #include <sycl/id.hpp>                                // for id
 #include <sycl/multi_ptr.hpp>                         // for multi_ptr
 #include <sycl/pointers.hpp>                          // for local_ptr, glo...
