+++ sycl/test/include_deps/sycl_khr_includes_event.hpp.cpp
@@ -54,5 +54,4 @@
 // CHECK-NEXT: detail/owner_less_base.hpp
 // CHECK-NEXT: detail/impl_utils.hpp
 // CHECK-NEXT: __spirv/spirv_vars.hpp
-// CHECK-NEXT: ext/oneapi/weak_object_base.hpp
 // CHECK-EMPTY:
