+++ sycl/test/include_deps/sycl_khr_includes_kernel_bundle.hpp.cpp
@@ -54,7 +54,6 @@
 // CHECK-NEXT: detail/owner_less_base.hpp
 // CHECK-NEXT: detail/impl_utils.hpp
 // CHECK-NEXT: __spirv/spirv_vars.hpp
-// CHECK-NEXT: ext/oneapi/weak_object_base.hpp
 // CHECK-NEXT: detail/util.hpp
 // CHECK-NEXT: detail/string.hpp
 // CHECK-NEXT: kernel_bundle_enums.hpp
